FIX - Bug with handling nils in the group-by keys.
This commit resolves an issue discovered when grouping by keys that
may involves nil values.   When `@groups.keys[0][1]` happens to be
a nil value for a key, then multi_indexed_grouping? would return false,
even though it should have returned true.  This commit changes
the test to check the size of grouping keys.